Utilize pom inheritance to simplify example poms.
A parent pom was created in pdfexamples and all pom configurations
common to all examples were moved from the children poms to the parent.
Furthermore the parent pom configures the maven-dependency-plugin to
copy dependencies to a common lib directory in pdfexamples since all
dependencies are shared. Various paths had to be edited to reflect this
change.