Move lib directory from project root to module/target.
The coc interpreter, for whatever reason, only is able to identify the
source code contained in the jars of the lib directory if the directory
is located in the modules target folder and not if the lib directory is
in the project root. For the sake of convenience the lib directory was
moved back to the individual module folders. This should be rectified in
the future to avoid duplication of the jars in  the project.